feat: add outbound HTTP support for TinyGo SDK
This commit adds support for sending HTTP requests from TinyGo components.

On the Spin implementation side, in addition to the existing outbound
HTTP library (https://github.com/deislabs/wasi-experimental-http),
this adds a new interface based on the WebAssembly component model.
This allows us to use the C tooling to generate bindings, then interop
with that using CGO.

The TinyGo API tries to follow the Go standard library as closely as
possible, and adds utility functions to send GET and POST requests,
as well as a function that sends an `*http.Request` (from `net/http`).

The CGO implementation is currently brittle, and should be heavily
improved in future commits, but the current state can be used to build
components that need to perform outbound HTTP requests.

Signed-off-by: Radu Matei <radu.matei@fermyon.com>

Co-authored-by: Radu Matei <radu.matei@fermyon.com>
Co-authored-by: Brian Ketelsen <bketelsen@gmail.com>